refactor: disable step creation on handle click
