+++ apps/journeys-admin/src/components/Editor/Slider/JourneyFlow/JourneyFlow.tsx
@@ -272,30 +272,35 @@ export function JourneyFlow(): ReactElement {
 
       if (step == null || block == null) return
 
-      let clientX = 0
-      let clientY = 0
-      if (event.type === 'touchend') {
-        const touchEvent = event as unknown as TouchEvent
-        touchEvent.preventDefault()
-        clientX = touchEvent.changedTouches[0].clientX
-        clientY = touchEvent.changedTouches[0].clientY
-      } else if (event.type === 'mouseup') {
-        const mouseEvent = event as unknown as MouseEvent
-        clientX = mouseEvent.clientX
-        clientY = mouseEvent.clientY
-      }
+      // let clientX = 0
+      // let clientY = 0
+      // if (event.type === 'touchend') {
+      //   const touchEvent = event as unknown as TouchEvent
+      //   touchEvent.preventDefault()
+      //   clientX = touchEvent.changedTouches[0].clientX
+      //   clientY = touchEvent.changedTouches[0].clientY
+      // } else if (event.type === 'mouseup') {
+      //   const mouseEvent = event as unknown as MouseEvent
+      //   clientX = mouseEvent.clientX
+      //   clientY = mouseEvent.clientY
+      // }
 
-      const { x, y } = reactFlowInstance.screenToFlowPosition({
-        x: clientX,
-        y: clientY
-      })
-
-      void createStepAndCardBlock(
-        step,
-        block,
-        parseInt(x.toString()),
-        parseInt(y.toString()) - STEP_NODE_HEIGHT / 2
+      const targetIsPane = (event.target as Element)?.classList.contains(
+        'react-flow__pane'
       )
+      if (targetIsPane) {
+        const { x, y } = reactFlowInstance.screenToFlowPosition({
+          x: (event as unknown as MouseEvent).clientX,
+          y: (event as unknown as MouseEvent).clientY
+        })
+
+        void createStepAndCardBlock(
+          step,
+          block,
+          parseInt(x.toString()),
+          parseInt(y.toString()) - STEP_NODE_HEIGHT / 2
+        )
+      }
     },
     [reactFlowInstance, connectingParams, createStepAndCardBlock, steps]
   )
