+++ apps/journeys-admin/src/components/Editor/Slider/JourneyFlow/nodes/StepBlockNode/StepBlockNode.tsx
@@ -4,7 +4,6 @@ import { NodeProps, OnConnect } from 'reactflow'
 
 import {
   ActiveContent,
-  ActiveFab,
   useEditor
 } from '@core/journeys/ui/EditorProvider'
 import { useJourney } from '@core/journeys/ui/JourneyProvider'
@@ -24,8 +23,7 @@ export const STEP_NODE_HEIGHT_GAP = 150
 
 export function StepBlockNode({ id, xPos, yPos }: NodeProps): ReactElement {
   const {
-    state: { steps, selectedStep, activeContent },
-    dispatch
+    state: { steps, selectedStep, activeContent }
   } = useEditor()
   const step = steps?.find((step) => step.id === id)
   const actionBlocks = filterActionBlocks(step)
@@ -57,22 +55,6 @@ export function StepBlockNode({ id, xPos, yPos }: NodeProps): ReactElement {
     })
   }
 
-  function handleClick(): void {
-    if (selectedStep?.id === step?.id) {
-      dispatch({
-        type: 'SetSelectedBlockAction',
-        selectedBlock: selectedStep
-      })
-      dispatch({ type: 'SetActiveFabAction', activeFab: ActiveFab.Add })
-      dispatch({
-        type: 'SetSelectedAttributeIdAction',
-        selectedAttributeId: `${selectedStep?.id ?? ''}-next-block`
-      })
-    } else {
-      dispatch({ type: 'SetSelectedStepAction', selectedStep: step })
-    }
-  }
-
   const isSelected =
     activeContent === ActiveContent.Canvas && selectedStep?.id === step?.id
 
@@ -89,7 +71,6 @@ export function StepBlockNode({ id, xPos, yPos }: NodeProps): ReactElement {
       <Stack
         gap={0.5}
         direction="column"
-        onClick={handleClick}
         sx={{
           background:
             activeContent === ActiveContent.Canvas &&
